requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
frozenlist==1.4.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
pip==24.2
propcache==0.2.0
pydantic-core==2.23.4
pydantic==2.9.2
python-dotenv==1.0.1
setuptools==65.5.0
typing-extensions==4.12.2
yarl==1.14.0
mistralai==1.1.0
SQLAlchemy==2.0.35
pytest==7.4.0